Days participation. Change to double amount fields.

+++ ConferenceAdmin/NancyService/Models/paymentbill.cs
@@ -8,7 +8,7 @@ namespace NancyService.Models
         public long paymentBillID { get; set; }
         public Nullable<long> paymentID { get; set; }
         public string transactionid { get; set; }
-        public int AmountPaid { get; set; }
+        public double AmountPaid { get; set; }
         public string methodOfPayment { get; set; }
         public string creditCardNumber { get; set; }
         public virtual payment payment { get; set; }
